requirements.txt
apiclient==1.0.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==4.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.32.3
rsa==4.6
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
wincertstore==0.2
PyPDF2
python-docx
scikit-learn
joblib
numpy
gunicorn
nltk==3.6.2
Django==5.2.1
django-heroku==0.3.1
dj-database-url==2.3.0
python-dotenv==1.0.0
whitenoise==6.9.0
psycopg2==2.9.10
reportlab==4.2.5
datasets==3.2.0
scikit-learn==1.5.2
joblib==1.4.2
